Switching to pre-baked postgres
