+++ crates/module-system/hyperlane/tests/integration/with_agent/helpers/docker.rs
@@ -1,24 +1,7 @@
 use std::time::Duration;
 use testcontainers::core::client::docker_client_instance;
 use testcontainers::core::ExecResult;
-use testcontainers::ContainerAsync;
 use tokio::io::AsyncBufReadExt;
-
-pub async fn print_logs_from_container<T>(name: &str, container: &ContainerAsync<T>)
-where
-    T: testcontainers::Image,
-{
-    let _span = tracing::info_span!("docker_log", name = name).entered();
-    let mut stdout = container.stdout(false).lines();
-    while let Some(line) = stdout.next_line().await.unwrap() {
-        tracing::info!("stdout: {line}");
-    }
-
-    let mut stderr = container.stderr(false).lines();
-    while let Some(line) = stderr.next_line().await.unwrap() {
-        tracing::info!("stderr: {line}");
-    }
-}
 
 pub async fn print_logs_from_exec_result(name: &str, result: &mut ExecResult, timeout: Duration) {
     let _span = tracing::info_span!("docker_exec_log", name = name).entered();
