+++ crates/module-system/sov-test-utils/Cargo.toml
@@ -27,6 +27,7 @@ serde = { workspace = true }
 schemars = { workspace = true }
 serde_json = { workspace = true }
 testcontainers = { workspace = true }
+testcontainers-modules = { workspace = true, features = ["postgres"] }
 tracing = { workspace = true }
 hex = { workspace = true }
 num_cpus = { workspace = true }
