+++ crates/module-system/sov-test-utils/src/docker.rs
@@ -1,0 +1,20 @@
+//! Docker related test-utils
+use testcontainers::ContainerAsync;
+use tokio::io::AsyncBufReadExt;
+
+/// Printing logs from container
+pub async fn print_logs_from_container<T>(name: &str, container: &ContainerAsync<T>)
+where
+    T: testcontainers::Image,
+{
+    let _span = tracing::info_span!("docker_log", name = name).entered();
+    let mut stdout = container.stdout(false).lines();
+    while let Some(line) = stdout.next_line().await.unwrap() {
+        tracing::info!("stdout: {line}");
+    }
+
+    let mut stderr = container.stderr(false).lines();
+    while let Some(line) = stderr.next_line().await.unwrap() {
+        tracing::info!("stderr: {line}");
+    }
+}
