+++ crates/module-system/sov-test-utils/src/postgres.rs
@@ -1,40 +1,7 @@
 use sov_sequencer::preferred::PostgresConfig;
-use std::borrow::Cow;
-use testcontainers::core::WaitFor;
 use testcontainers::runners::AsyncRunner;
-use testcontainers::{ContainerAsync, Image};
-
-/// A Docker image for PostgreSQL.
-#[derive(Debug, Clone, Default)]
-pub struct PostgresImage;
-
-impl Image for PostgresImage {
-    fn name(&self) -> &str {
-        "postgres"
-    }
-
-    fn tag(&self) -> &str {
-        "17-alpine"
-    }
-
-    fn ready_conditions(&self) -> Vec<WaitFor> {
-        // See <https://github.com/testcontainers/testcontainers-rs-modules-community/issues/158>.
-        vec![
-            WaitFor::message_on_stderr("database system is ready to accept connections"),
-            WaitFor::message_on_stdout("database system is ready to accept connections"),
-        ]
-    }
-
-    fn env_vars(
-        &self,
-    ) -> impl IntoIterator<Item = (impl Into<Cow<'_, str>>, impl Into<Cow<'_, str>>)> {
-        [
-            ("POSTGRES_DB", "postgres"),
-            ("POSTGRES_USER", "postgres"),
-            ("POSTGRES_PASSWORD", "postgres"),
-        ]
-    }
-}
+use testcontainers::{ContainerAsync, ImageExt};
+use testcontainers_modules::postgres::Postgres;
 
 #[derive(Debug, thiserror::Error)]
 /// Error indicating problems when creating a Postgres container.
@@ -49,13 +16,13 @@ pub enum CreatePostgresError {
 }
 
 /// Creates a container with a PostgreSQL database.
-pub async fn create_postgres_container(
-) -> Result<ContainerAsync<PostgresImage>, CreatePostgresError> {
+pub async fn create_postgres_container() -> Result<ContainerAsync<Postgres>, CreatePostgresError> {
     if should_skip_postgres() {
         return Err(CreatePostgresError::DockerNotSupported);
     }
 
-    let img = PostgresImage
+    let img = Postgres::default()
+        .with_tag("17-alpine")
         .start()
         .await
         .map_err(|e| CreatePostgresError::DockerError(e.into()))?;
@@ -83,7 +50,7 @@ fn should_skip_postgres() -> bool {
 
 /// Returns the connection string for the PostgreSQL.
 pub async fn connection_string_from_postgres_container(
-    container: &ContainerAsync<PostgresImage>,
+    container: &ContainerAsync<Postgres>,
 ) -> anyhow::Result<String> {
     let postgres_connection_string = format!(
         "postgres://postgres:postgres@{}:{}",
@@ -96,7 +63,7 @@ pub async fn connection_string_from_postgres_container(
 
 /// Returns the connection string for the PostgreSQL.
 pub async fn config_from_postgres_container(
-    container: &ContainerAsync<PostgresImage>,
+    container: &ContainerAsync<Postgres>,
     node_id: String,
 ) -> anyhow::Result<PostgresConfig> {
     let postgres_connection_string = connection_string_from_postgres_container(container).await?;
